+++ index.js
@@ -2,7 +2,11 @@
 'use strict'
 
 const Forms = require('./lib/forms.js')
+const Keys = require('./lib/keys.js')
+const Organisations = require('./lib/organisations.js')
 
 module.exports = {
-  Forms
+  Forms,
+  Keys,
+  Organisations
 }